refactor(canvas): deepen persistence internals
