+++ apps/canvas-workspace/src/main/canvas/storage.ts
@@ -6,17 +6,16 @@
  * the (separately-packaged) `canvas-cli`. No Electron imports here so the
  * module is unit-testable in plain Node.
  *
- * Today: provides one atomic write implementation, recovery-aware reads, and
- * (dormant) v2 split-storage primitives that are not yet wired up. PR1
- * lands the helpers and tests; PR3 flips on lazy auto-migration once the
- * other consumers are also routing through here.
+ * This file is the stable compatibility facade for full-canvas reads, writes,
+ * and v1→v2 migration. Owner-local path, atomic JSON, schema, and pollution
+ * behavior lives under `persistence/`; callers continue importing this file.
  *
  * v1 vs v2 storage layout:
  *
- *   v1 (current):
+ *   v1 (legacy, migrated lazily):
  *     ~/.pulse-coder/canvas/<id>/canvas.json   ← layout + ALL node.data inline
  *
- *   v2 (target):
+ *   v2 (current):
  *     ~/.pulse-coder/canvas/<id>/canvas.json   ← layout-only (no node.data)
  *     ~/.pulse-coder/canvas/<id>/nodes/<nodeId>.json
  *                                              ← self-describing per-node file
@@ -24,508 +23,96 @@
  * Migration is workspace-scoped and atomic at the canvas.json swap.
  */
 
-import type { Dirent } from 'fs';
 import { promises as fs } from 'fs';
-import { join, basename, dirname } from 'path';
-import { homedir } from 'os';
+import { dirname } from 'path';
 import {
-  WORKSPACE_NODE_SCHEMA_VERSION,
-  assertSafeNodeId as assertSafeWorkspaceNodeId,
   deleteWorkspaceNode,
-  getNodeFilePath as getWorkspaceNodeFilePath,
-  getNodesDir as getWorkspaceNodesDir,
-  isSafeNodeId as isSafeWorkspaceNodeId,
   listWorkspaceNodeIds,
   mutateWorkspaceNode,
   readWorkspaceNode,
   writeWorkspaceNode,
-  type WorkspaceNodeLink,
-  type WorkspaceNodePropertyValue,
-  type WorkspaceNodeRecord,
 } from './nodes/store';
+import {
+  CANVAS_JSON_FILENAME,
+  MANIFEST_ID,
+  MIGRATION_SENTINEL_FILENAME,
+  NODES_DIR_NAME,
+  STORE_DIR,
+  V1_BACKUP_FILENAME,
+  assertSafeNodeId,
+  getCanvasJsonPath,
+  getNodeFilePath,
+  getNodesDir,
+  getSentinelPath,
+  getV1BackupPath,
+  getV1TimestampedBackupPath,
+  getWorkspaceDir,
+  isSafeNodeId,
+} from './persistence/paths';
+import {
+  atomicWriteJson,
+  isEnoent,
+  readJsonWithRecovery,
+  type ReadJsonResult,
+} from './persistence/atomic-json';
+import {
+  CANVAS_SCHEMA_VERSION_V2,
+  PER_NODE_SCHEMA_VERSION,
+  detectSchemaVersion,
+  type CanvasNode,
+  type CanvasSaveData,
+  type MigrationProgress,
+  type MigrationSentinel,
+  type PerNodeFile,
+  type ReadCanvasResult,
+  type SchemaVersion,
+} from './persistence/schema';
+import {
+  CanvasPollutionDetectedError,
+  detectV1Pollution,
+  scanForPollutedWorkspaces,
+} from './persistence/pollution';
 
-// ─────────────────────────────────────────────────────────────────────────────
-// Constants
-
-/** Root for all per-workspace storage. Mirrors `canvas-store.ts`. */
-export const STORE_DIR = join(homedir(), '.pulse-coder', 'canvas');
-
-/** Manifest id — flat file at STORE_DIR root, not a workspace directory. */
-export const MANIFEST_ID = '__workspaces__';
-
-export const CANVAS_JSON_FILENAME = 'canvas.json';
-export const NODES_DIR_NAME = 'nodes';
-/** Permanent v1 archive, written once at migration start. Never overwritten. */
-export const V1_BACKUP_FILENAME = 'canvas.json.v1.bak';
-/** Sentinel; present iff a migration is mid-flight. */
-export const MIGRATION_SENTINEL_FILENAME = '.migrating';
-
-/** Current per-node file schema version. */
-export const PER_NODE_SCHEMA_VERSION = WORKSPACE_NODE_SCHEMA_VERSION;
-/** Target canvas.json schema version once migration completes. */
-export const CANVAS_SCHEMA_VERSION_V2 = 2;
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Types
-
-/**
- * A single canvas node, v1-shape — what callers and the renderer always work
- * with in memory. `data` is the polymorphic payload that varies by `type`.
- * In v2 on-disk format, `data` lives in `nodes/<id>.json` instead of being
- * inlined here, but assembled reads still hand callers this shape.
- *
- * Intentionally NO `[k: string]: unknown` index signature: consumers
- * (canvas-store, canvas-agent/tools, mcp-server, etc.) have their own
- * strictly-typed `CanvasNode` interfaces and a wildcard signature here
- * would block structural assignment from them. The helper itself only
- * reads named fields, so the wider shape isn't needed.
- */
-export interface CanvasNode {
-  id?: string;
-  type: string;
-  title?: string;
-  x?: number;
-  y?: number;
-  width?: number;
-  height?: number;
-  ref?: unknown;
-  data?: Record<string, unknown>;
-  properties?: Record<string, WorkspaceNodePropertyValue>;
-  links?: WorkspaceNodeLink[];
-  /** Epoch millis of last mutation; used for cross-process merge. */
-  updatedAt?: number;
-}
-
-/**
- * The full v1-shape canvas data. Returned by `readCanvasFull` regardless of
- * on-disk schema version, accepted by `writeCanvasFull`. `schemaVersion` is
- * the *target* version that should end up on disk after a write; callers
- * don't need to set it (defaults to whatever is currently on disk, or v2 for
- * fresh workspaces).
- */
-export interface CanvasSaveData {
-  schemaVersion?: 1 | 2;
-  nodes?: CanvasNode[];
-  edges?: unknown[];
-  transform?: unknown;
-  savedAt?: string;
-}
-
-/** Back-compat alias for the workspace-local atomic node record. */
-export type PerNodeFile = WorkspaceNodeRecord;
-export type { WorkspaceNodeLink, WorkspaceNodePropertyValue, WorkspaceNodeRecord };
-
-export type SchemaVersion = 1 | 2;
-
-/**
- * Sentinel content written when migration starts. The `expectedNodeIds` list
- * lets recovery cleanup distinguish "partial migration garbage" from user
- * files that happen to live in the nodes/ directory.
- */
-export interface MigrationSentinel {
-  startedAt: number;
-  workspaceId: string;
-  sourceUpdatedAt: number | null;
-  expectedNodeIds: string[];
-}
-
-/**
- * Progress callback payload for `migrateToV2`. Phases run in order; total +
- * current are only meaningful during `split-nodes`. `error` is not emitted
- * by `migrateToV2` itself (it throws on failure) but is included so the
- * canvas-store-side broadcaster can publish error events on the same
- * channel as progress events without a separate type.
- */
-export interface MigrationProgress {
-  phase: 'starting' | 'backup' | 'split-nodes' | 'commit' | 'done' | 'error';
-  current?: number;
-  total?: number;
-  message?: string;
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Path helpers
-
-/** Workspace directory; the layout root for everything else. */
-export function getWorkspaceDir(workspaceId: string, root: string = STORE_DIR): string {
-  return join(root, workspaceId);
-}
-
-/** Path to a workspace's `canvas.json`. Manifest is a flat file at STORE_DIR root. */
-export function getCanvasJsonPath(workspaceId: string, root: string = STORE_DIR): string {
-  if (workspaceId === MANIFEST_ID) {
-    return join(root, `${MANIFEST_ID}.json`);
-  }
-  return join(getWorkspaceDir(workspaceId, root), CANVAS_JSON_FILENAME);
-}
-
-export function getNodesDir(workspaceId: string, root: string = STORE_DIR): string {
-  return getWorkspaceNodesDir(workspaceId, root);
-}
-
-/**
- * Path to a single per-node file. Throws on suspicious node ids to keep
- * path-traversal out of writers — node ids in canvas.json come from disk
- * and could be tampered with.
- */
-export function getNodeFilePath(workspaceId: string, nodeId: string, root: string = STORE_DIR): string {
-  return getWorkspaceNodeFilePath(workspaceId, nodeId, root);
-}
-
-export function getV1BackupPath(workspaceId: string, root: string = STORE_DIR): string {
-  return join(getWorkspaceDir(workspaceId, root), V1_BACKUP_FILENAME);
-}
-
-/**
- * Path to a timestamped immutable v1 archive. Each migration writes a new
- * one of these and never overwrites prior files, so the original
- * pre-v2 snapshot survives even if a later (e.g., pollution-triggered)
- * re-migration overwrites the stable `canvas.json.v1.bak` alias.
- *
- * Naming: `canvas.json.v1.<ISO-UTC-stamp>.bak`, with `:`/`.` swapped to
- * `-` so the filename is filesystem-friendly on every OS.
- */
-export function getV1TimestampedBackupPath(
-  workspaceId: string,
-  timestamp: Date = new Date(),
-  root: string = STORE_DIR,
-): string {
-  const stamp = timestamp.toISOString().replace(/[:.]/g, '-');
-  return join(
-    getWorkspaceDir(workspaceId, root),
-    `canvas.json.v1.${stamp}.bak`,
-  );
-}
-
-export function getSentinelPath(workspaceId: string, root: string = STORE_DIR): string {
-  return join(getWorkspaceDir(workspaceId, root), MIGRATION_SENTINEL_FILENAME);
-}
-
-/**
- * Node-id whitelist. Rejects path traversal, separators, and shell-style
- * special chars. Matches the generator in renderer (`node-<ts>-<n>`) and
- * future genId variants while staying conservative.
- */
-export function isSafeNodeId(id: string): boolean {
-  return isSafeWorkspaceNodeId(id);
-}
-
-export function assertSafeNodeId(id: string): void {
-  assertSafeWorkspaceNodeId(id);
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Atomic I/O
-
-function isEnoent(err: unknown): boolean {
-  return !!err && typeof err === 'object' && (err as { code?: string }).code === 'ENOENT';
-}
-
-function shouldRotateRollingBackup(parsed: unknown): boolean {
-  if (!parsed || typeof parsed !== 'object') return false;
-  const obj = parsed as { nodes?: unknown[]; workspaces?: unknown[]; entries?: unknown[] };
-  return (
-    (Array.isArray(obj.nodes) && obj.nodes.length > 0) ||
-    (Array.isArray(obj.workspaces) && obj.workspaces.length > 0) ||
-    (Array.isArray(obj.entries) && obj.entries.length > 0)
-  );
-}
-
-/**
- * Atomically write JSON to disk via tmp + rename.
- *
- * `rename()` is the only safe way to publish a new file: `fs.writeFile` is
- * NOT atomic (truncate-then-stream), so a crash or a concurrent reader can
- * see an empty/partial file. With multiple writers racing on the same file
- * (canvas-workspace + canvas-cli + canvas-agent + MCP server, all on the
- * same workspace) this is a real data-loss path.
- *
- * If `rollingBackup` is true, the current contents are copied to
- * `<path>.bak` *before* overwrite — but only when the current file parses
- * and looks non-empty. That prevents a single corrupt write from
- * poisoning the rolling backup.
- */
-export async function atomicWriteJson(
-  finalPath: string,
-  serialized: string,
-  opts: { rollingBackup?: boolean } = {},
-): Promise<void> {
-  const dir = dirname(finalPath);
-  const base = basename(finalPath);
-  const tmpPath = join(
-    dir,
-    `${base}.${process.pid}.${Date.now()}.${Math.random().toString(36).slice(2, 8)}.tmp`,
-  );
-  const bakPath = join(dir, `${base}.bak`);
-
-  await fs.mkdir(dir, { recursive: true });
-  await fs.writeFile(tmpPath, serialized, 'utf-8');
-
-  if (opts.rollingBackup) {
-    try {
-      const currentRaw = await fs.readFile(finalPath, 'utf-8');
-      try {
-        // Only rotate when the current file is a *good* snapshot. Don't let
-        // a corrupt save poison the last-known-good backup.
-        if (shouldRotateRollingBackup(JSON.parse(currentRaw))) {
-          await fs.copyFile(finalPath, bakPath).catch(() => undefined);
-        }
-      } catch {
-        // Current file unparseable — keep existing .bak intact.
-      }
-    } catch {
-      // No current file (first write) — nothing to back up.
-    }
-  }
-
-  await fs.rename(tmpPath, finalPath);
-}
-
-/**
- * Result of a recovery-aware JSON read.
- *  - ok:           parsed cleanly (from primary or backup).
- *  - missing:      neither file exists; caller treats as fresh.
- *  - unrecoverable: primary failed AND backup failed; caller must surface.
- */
-export type ReadJsonResult<T = unknown> =
-  | { kind: 'ok'; data: T; recoveredFromBackup: boolean }
-  | { kind: 'missing' }
-  | { kind: 'unrecoverable'; err: unknown };
-
-/**
- * Read JSON with transparent fallback to `<path>.bak`. Used for `canvas.json`
- * so that even an older, pre-atomic-write corruption can self-heal on the
- * next load.
- */
-export async function readJsonWithRecovery<T = unknown>(
-  finalPath: string,
-): Promise<ReadJsonResult<T>> {
-  const bakPath = `${finalPath}.bak`;
-  let primaryErr: unknown = null;
-
-  try {
-    const raw = await fs.readFile(finalPath, 'utf-8');
-    try {
-      return { kind: 'ok', data: JSON.parse(raw) as T, recoveredFromBackup: false };
-    } catch (err) {
-      primaryErr = err;
-    }
-  } catch (err) {
-    if (isEnoent(err)) {
-      // Fall through to backup; if also missing, return `missing`.
-    } else {
-      primaryErr = err;
-    }
-  }
-
-  try {
-    const bakRaw = await fs.readFile(bakPath, 'utf-8');
-    return {
-      kind: 'ok',
-      data: JSON.parse(bakRaw) as T,
-      recoveredFromBackup: true,
-    };
-  } catch (bakErr) {
-    if (isEnoent(bakErr)) {
-      return primaryErr
-        ? { kind: 'unrecoverable', err: primaryErr }
-        : { kind: 'missing' };
-    }
-    return { kind: 'unrecoverable', err: primaryErr ?? bakErr };
-  }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Schema detection
-
-/**
- * Detect on-disk schema version. v1 is identified by `schemaVersion` either
- * absent, undefined, or strictly `1`. Anything else with `nodes` is treated
- * as v1 too (forward-tolerant: an unknown future version we can still read
- * by ignoring extra fields).
- */
-export function detectSchemaVersion(parsed: unknown): SchemaVersion {
-  if (parsed && typeof parsed === 'object') {
-    const v = (parsed as { schemaVersion?: unknown }).schemaVersion;
-    if (v === 2) return 2;
-  }
-  return 1;
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Pollution detection
-//
-// Background: a v1-unaware writer (old binary, external script that does
-// fs.readFile + JSON.parse + fs.writeFile on canvas.json) that touches a
-// v2 workspace will produce a v1-shape canvas.json — node entries without
-// `data` fields and without `schemaVersion: 2`. By itself that's just a
-// stale canvas.json; the real data is still safe in nodes/<id>.json. But
-// the next v2 read would treat it as a legitimate v1 workspace and trigger
-// `migrateToV2`, which would copy the empty-data v1 layout into per-node
-// files (since updatedAt arbitration almost always lets incoming win) and
-// permanently destroy the original data. This guard refuses both the
-// v1-shape write and the migration once the signature is present.
-
-/**
- * Thrown when a write or migration that would normally produce a v1-shape
- * canvas.json detects existing v2 per-node files for the same workspace.
- * The fact that nodes/<id>.json exists for ids that the incoming
- * v1-shape canvas mentions is the smoking gun: this workspace was
- * previously v2, and a v1-unaware code path is now trying to clobber it.
- */
-export class CanvasPollutionDetectedError extends Error {
-  readonly workspaceId: string;
-  readonly conflictingNodeIds: string[];
-  constructor(workspaceId: string, conflictingNodeIds: string[]) {
-    const sample = conflictingNodeIds.slice(0, 5).join(', ');
-    const more =
-      conflictingNodeIds.length > 5
-        ? `, +${conflictingNodeIds.length - 5} more`
-        : '';
-    super(
-      `[canvas-storage] refusing v1-shape write/migration for workspace ` +
-      `"${workspaceId}": ${conflictingNodeIds.length} node id(s) already ` +
-      `have v2 per-node files on disk (${sample}${more}). This is the ` +
-      `signature of a v1-unaware writer (old binary or external script) ` +
-      `having clobbered canvas.json. The real data is still in the ` +
-      `nodes/<id>.json files — do NOT migrate, restore canvas.json ` +
-      `instead (see docs).`,
-    );
-    this.name = 'CanvasPollutionDetectedError';
-    this.workspaceId = workspaceId;
-    this.conflictingNodeIds = conflictingNodeIds;
-  }
-}
-
-/**
- * For each node id mentioned in the incoming v1-shape canvas, return the
- * subset whose on-disk state matches the v1-pollution signature.
- *
- * The signature we're catching: a v1-unaware writer reads v2 canvas.json
- * (layout-only, no `data` per node) and writes it back verbatim. The
- * per-node files in `nodes/<id>.json` still hold the real content, and
- * letting migration proceed in that state would let `updatedAt`
- * arbitration overwrite each per-node file with the incoming v1 layout
- * and permanently destroy the user's data.
- *
- * We require BOTH conditions per node to call it pollution:
- *
- *   1. The on-disk `nodes/<id>.json` actually carries v2 *content* — a
- *      non-empty `data` object or non-empty `links`. A per-node file
- *      that only holds `properties` (e.g. tags written by
- *      `canvas_tag_node` against an otherwise-v1 workspace) is
- *      metadata side-data that legitimately coexists with a v1
- *      canvas.json and must NOT be treated as pollution — that path
- *      was producing false positives that masked the real signal.
- *
- *   2. The incoming v1 canvas node has empty or missing `data` — the
- *      smoking gun of a v1-unaware writer that read the stripped v2
- *      layout and wrote it back as v1. A v1 node that still carries
- *      real inline `data` is just a normal v1 workspace (or an export
- *      from a v2 reader that re-inlined data), and nothing was lost.
- *
- * Either condition alone is insufficient: condition 1 alone misfires on
- * workspaces where the agent wrote tags but the canvas itself is still
- * a healthy v1 (the bug this comment replaces); condition 2 alone
- * misfires on legitimate v1 workspaces that happen to have empty `data`
- * for layout-only node types (frames, groupings).
- *
- * Reads the per-node files (not just fs.access). That's still cheap —
- * each file is small JSON and only conflicting *candidates* are read.
- */
-export async function detectV1Pollution(
-  workspaceId: string,
-  incomingNodes: CanvasNode[] | undefined,
-  root: string = STORE_DIR,
-): Promise<string[]> {
-  if (!Array.isArray(incomingNodes) || incomingNodes.length === 0) return [];
-
-  const byId = new Map<string, CanvasNode>();
-  for (const n of incomingNodes) {
-    if (typeof n.id === 'string' && isSafeNodeId(n.id)) byId.set(n.id, n);
-  }
-  if (byId.size === 0) return [];
-
-  const dataNonEmpty = (d: unknown): boolean =>
-    !!d && typeof d === 'object' && !Array.isArray(d) && Object.keys(d as object).length > 0;
-
-  const conflicts: string[] = [];
-  await Promise.all(
-    [...byId.entries()].map(async ([id, incoming]) => {
-      // Condition 2 first — cheap, no I/O. If the incoming v1 node still
-      // carries real inline `data`, no data is at risk: bail.
-      if (dataNonEmpty((incoming as { data?: unknown }).data)) return;
-
-      // Condition 1 — read the per-node file. Missing/unreadable is
-      // not pollution (matches the legacy fs.access semantics).
-      let onDisk: WorkspaceNodeRecord;
-      try {
-        const raw = await fs.readFile(getNodeFilePath(workspaceId, id, root), 'utf-8');
-        onDisk = JSON.parse(raw) as WorkspaceNodeRecord;
-      } catch {
-        return;
-      }
-      const linksNonEmpty = Array.isArray(onDisk.links) && onDisk.links.length > 0;
-      if (!dataNonEmpty(onDisk.data) && !linksNonEmpty) return;
-
-      conflicts.push(id);
-    }),
-  );
-  return conflicts;
-}
-
-/**
- * Walk every workspace directory under `root` and return those whose
- * on-disk state matches the v1-pollution signature (canvas.json is v1
- * shape AND at least one of its node ids has a corresponding
- * nodes/<id>.json file).
- *
- * Used at app startup so the renderer can surface a sticky alert for
- * each affected workspace before the user clicks one — surfacing only
- * on open would mean the user has to actually try to interact with the
- * polluted workspace to learn it's broken, by which point the
- * canvas:save guard would refuse but the spinner's error toast might
- * be missed.
- *
- * Cheap: skips workspaces whose canvas.json already says v2 (the
- * common case) before reaching for any per-node files.
- */
-export async function scanForPollutedWorkspaces(
-  root: string = STORE_DIR,
-): Promise<Array<{ workspaceId: string; conflictingNodeIds: string[] }>> {
-  let entries: Dirent[];
-  try {
-    entries = (await fs.readdir(root, { withFileTypes: true })) as Dirent[];
-  } catch (err) {
-    if (isEnoent(err)) return [];
-    throw err;
-  }
-
-  const findings: Array<{ workspaceId: string; conflictingNodeIds: string[] }> = [];
-  for (const entry of entries) {
-    if (!entry.isDirectory() || entry.name === MANIFEST_ID) continue;
-    const workspaceId = entry.name;
-    const canvasPath = getCanvasJsonPath(workspaceId, root);
-
-    let parsed: CanvasSaveData;
-    try {
-      const raw = await fs.readFile(canvasPath, 'utf-8');
-      parsed = JSON.parse(raw) as CanvasSaveData;
-    } catch {
-      // Workspace with missing or unreadable canvas.json — not the
-      // pollution signature; skip.
-      continue;
-    }
-
-    if (detectSchemaVersion(parsed) === 2) continue;
-
-    const conflicts = await detectV1Pollution(workspaceId, parsed.nodes, root);
-    if (conflicts.length > 0) {
-      findings.push({ workspaceId, conflictingNodeIds: conflicts });
-    }
-  }
-  return findings;
-}
+export {
+  CANVAS_JSON_FILENAME,
+  MANIFEST_ID,
+  MIGRATION_SENTINEL_FILENAME,
+  NODES_DIR_NAME,
+  STORE_DIR,
+  V1_BACKUP_FILENAME,
+  assertSafeNodeId,
+  atomicWriteJson,
+  getCanvasJsonPath,
+  getNodeFilePath,
+  getNodesDir,
+  getSentinelPath,
+  getV1BackupPath,
+  getV1TimestampedBackupPath,
+  getWorkspaceDir,
+  isSafeNodeId,
+  readJsonWithRecovery,
+};
+export type { ReadJsonResult };
+export {
+  CANVAS_SCHEMA_VERSION_V2,
+  PER_NODE_SCHEMA_VERSION,
+  CanvasPollutionDetectedError,
+  detectSchemaVersion,
+  detectV1Pollution,
+  scanForPollutedWorkspaces,
+};
+export type {
+  CanvasNode,
+  CanvasSaveData,
+  MigrationProgress,
+  MigrationSentinel,
+  PerNodeFile,
+  ReadCanvasResult,
+  SchemaVersion,
+  WorkspaceNodeLink,
+  WorkspaceNodePropertyValue,
+  WorkspaceNodeRecord,
+} from './persistence/schema';
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Sentinel
@@ -719,15 +306,6 @@ export async function listNodeFiles(
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Full read / write
-
-export interface ReadCanvasResult {
-  /** Full v1-shape data ready for callers; null if no canvas.json exists. */
-  data: CanvasSaveData | null;
-  /** True if the primary canvas.json was unreadable and `.bak` was used. */
-  recoveredFromBackup: boolean;
-  /** On-disk schema; null when the workspace has no canvas.json at all. */
-  schemaVersion: SchemaVersion | null;
-}
 
 /**
  * Read the workspace's canvas in v1-shape (with `node.data` inline),
